apps/podcast-agent/requirements.txt
gradio
google-genai
edge-tts
pypdf
anyio
numpy
librosa
soundfile
